feat: everything inside the terminal takes the fixed palette
zellij, btop, bat, delta, lazygit, opencode and herdr move into
terminal-surfaces.ts and stop reading a theme. Neovim theming is deleted
outright.

They were split across two files by an accident of history — zellij and
btop sat with the desktop writers, the rest in theme-cli.ts — and they
belong together now that they share the property that matters: none of
them varies. theme-cli.ts is gone; every function in it moved.

Two of these already worked this way. opencode's "theme": "system" and
herdr's name = "terminal" both mean *follow the host terminal*, and
neither ever read a theme; they were in a theme registry because that is
where the code happened to be.

bat drops its ten-slug mapping for base16, which renders through the
terminal's own ANSI sixteen — now fixed and RedDB's. The mapping it
replaces sent four of the ten themes to "Nord", which is a fair measure
of how much signal it carried. delta reads bat's themes, so one name
serves both.

The marker text changed, and two writers find their own previous output
by looking for it. They match the prefix rather than the sentence, or
every already-configured machine would report "not ours — left alone"
and never be updated again.

The zellij fallback in dotfiles.ts becomes unconditional. It was guarded
on the file being absent because the content depended on which theme was
chosen and a default would have clobbered a real one. There is one
palette now, so reasserting it is free and repairs a hand-edited file.

Verified on a real machine rather than in the log: the written kdl,
btop theme and bat config all carry ink #07080a, accent #ff2056 and
foreground #b3b8c4.
