+++ src/terminal-surfaces.ts
@@ -1,0 +1,383 @@
+/**
+ * Everything that renders inside the terminal window.
+ *
+ * These used to be theme surfaces, split across two files by an
+ * accident of history: zellij and btop lived with the desktop writers
+ * in theme-apply.ts, while bat, delta, lazygit, opencode and herdr lived
+ * in theme-cli.ts. They belong together, because they now share one
+ * property that matters more than where they were born — none of them
+ * varies with the theme.
+ *
+ * See .red/adr/0002. The short version: a program inside a terminal
+ * paints over the ANSI slots underneath it, so a theme spread across a
+ * dozen of them arrived as neither the old theme nor the new one. They
+ * all take the fixed RedDB palette now, and the themes moved to the
+ * wallpaper, the system accent, GNOME and VS Code — surfaces where a
+ * change is unambiguous.
+ *
+ * Two of these already worked this way before the change. opencode's
+ * `"theme": "system"` and herdr's `name = "terminal"` both mean *follow
+ * the host terminal*, which is the same answer omarchy reached. They
+ * were only ever in a theme registry because that is where the code
+ * happened to be; nothing about them read a theme.
+ *
+ * Each writer stays independent and failure-tolerant: a missing
+ * application is not an error, just a surface this machine does not
+ * have.
+ */
+
+import { existsSync, mkdirSync } from "node:fs";
+import { log } from "./log.ts";
+import { configHome } from "./shared-root.ts";
+import { RED_DEV_ANSI } from "./terminal-palette.ts";
+import type { Platform } from "./platform.ts";
+
+/**
+ * The marker, and why the prefix is matched separately.
+ *
+ * The old text said "overwritten on theme switch", which stopped being
+ * true. Two writers below find their own previous output by looking for
+ * this line, so changing the wording would make every already-configured
+ * machine report "not ours — left alone" and never be updated again.
+ * They match MARKER_PREFIX, which both spellings share.
+ */
+const GENERATED = "Generated by red-dev. The terminal palette does not change with the theme.";
+const MARKER_PREFIX = "Generated by red-dev";
+
+// ----------------------------------------------------------- zellij
+
+/**
+ * zellij reads themes from config.kdl or from a themes directory. We
+ * write a themes/ file and reference it, so the user's own config.kdl
+ * keeps whatever keybindings and layout settings they have set.
+ */
+export async function applyZellij(p: Platform): Promise<boolean> {
+  // configHome, not home(): when a shared root exists this writes into
+  // it, so what the distro writes is what Windows reads. Every writer
+  // here used to hardcode ~/.config, which is what made the share
+  // something you migrated into rather than something configuration was
+  // born in.
+  //
+  // btop below deliberately keeps home(): its config names an absolute
+  // path to its own theme file, so a shared copy would point one side at
+  // the other's home.
+  const dir = `${configHome(p, "zellij")}/zellij`;
+  const themesDir = `${dir}/themes`;
+  mkdirSync(themesDir, { recursive: true });
+
+  const c = RED_DEV_ANSI;
+  const kdl = `// ${GENERATED}
+themes {
+    red-dev {
+        fg "${c.foreground}"
+        bg "${c.background}"
+        black "${c.black}"
+        red "${c.red}"
+        green "${c.green}"
+        yellow "${c.yellow}"
+        blue "${c.blue}"
+        magenta "${c.purple}"
+        cyan "${c.cyan}"
+        white "${c.white}"
+        orange "${c.yellow}"
+    }
+}
+`;
+  await Bun.write(`${themesDir}/red-dev.kdl`, kdl);
+
+  // Point config.kdl at it without disturbing anything else in there.
+  const configPath = `${dir}/config.kdl`;
+  const existing = existsSync(configPath) ? await Bun.file(configPath).text() : "";
+
+  if (existing.includes('theme "red-dev"')) return true;
+
+  const withoutOldTheme = existing.replace(/^\s*theme\s+".*"\s*$/gm, "").trimEnd();
+  const next = `${withoutOldTheme}\n\n// ${GENERATED}\ntheme "red-dev"\n`;
+  await Bun.write(configPath, next.trimStart());
+  return true;
+}
+
+// ------------------------------------------------------------- btop
+
+export async function applyBtop(): Promise<boolean> {
+  const h = process.env["HOME"] ?? process.env["USERPROFILE"];
+  if (!h) throw new Error("neither HOME nor USERPROFILE is set");
+
+  const dir = `${h}/.config/btop`;
+  const themesDir = `${dir}/themes`;
+  mkdirSync(themesDir, { recursive: true });
+
+  const c = RED_DEV_ANSI;
+  // btop's format is flat key/value with quoted hex.
+  const btop = `# ${GENERATED}
+theme[main_bg]="${c.background}"
+theme[main_fg]="${c.foreground}"
+theme[title]="${c.foreground}"
+theme[hi_fg]="${c.blue}"
+theme[selected_bg]="${c.selectionBackground}"
+theme[selected_fg]="${c.foreground}"
+theme[inactive_fg]="${c.brightBlack}"
+theme[graph_text]="${c.foreground}"
+theme[proc_misc]="${c.green}"
+theme[cpu_box]="${c.blue}"
+theme[mem_box]="${c.green}"
+theme[net_box]="${c.purple}"
+theme[proc_box]="${c.cyan}"
+theme[div_line]="${c.brightBlack}"
+theme[temp_start]="${c.green}"
+theme[temp_mid]="${c.yellow}"
+theme[temp_end]="${c.red}"
+theme[cpu_start]="${c.green}"
+theme[cpu_mid]="${c.yellow}"
+theme[cpu_end]="${c.red}"
+theme[free_start]="${c.green}"
+theme[free_mid]="${c.cyan}"
+theme[free_end]="${c.blue}"
+theme[cached_start]="${c.cyan}"
+theme[cached_mid]="${c.blue}"
+theme[cached_end]="${c.purple}"
+theme[available_start]="${c.yellow}"
+theme[available_mid]="${c.yellow}"
+theme[available_end]="${c.red}"
+theme[used_start]="${c.green}"
+theme[used_mid]="${c.yellow}"
+theme[used_end]="${c.red}"
+theme[download_start]="${c.purple}"
+theme[download_mid]="${c.blue}"
+theme[download_end]="${c.cyan}"
+theme[upload_start]="${c.red}"
+theme[upload_mid]="${c.yellow}"
+theme[upload_end]="${c.green}"
+`;
+  await Bun.write(`${themesDir}/red-dev.theme`, btop);
+
+  // btop.conf is a flat file; rewrite only the color_theme line.
+  const confPath = `${dir}/btop.conf`;
+  const conf = existsSync(confPath) ? await Bun.file(confPath).text() : "";
+  const line = `color_theme = "${themesDir}/red-dev.theme"`;
+
+  if (conf.includes("color_theme")) {
+    await Bun.write(confPath, conf.replace(/^color_theme\s*=.*$/m, line));
+  } else {
+    await Bun.write(confPath, `${conf.trimEnd()}\n${line}\ntheme_background = False\n`.trimStart());
+  }
+  return true;
+}
+
+// --------------------------------------------------------- bat, delta
+
+/**
+ * The one bat theme, and why it is base16.
+ *
+ * bat ships a fixed set built from Sublime syntaxes, so there is no
+ * palette to emit — only a name to pick. `base16` renders through the
+ * terminal's own ANSI sixteen, which are now RedDB's and fixed. So bat
+ * follows the terminal for the same reason opencode and herdr do, rather
+ * than pinning a published theme that would disagree with it.
+ *
+ * The mapping this replaces sent four of the ten themes to "Nord",
+ * which is a fair summary of how much signal it carried.
+ */
+const BAT_THEME = "base16";
+
+/**
+ * bat, and the Debian rename that makes it two tools.
+ *
+ * Debian ships the binary as `batcat` because `bat` was taken, and the
+ * config directory follows the binary: `~/.config/batcat` there,
+ * `~/.config/bat` everywhere else. Writing only one leaves the tool
+ * reading a file that does not exist and reporting nothing wrong.
+ */
+export async function applyBat(p: Platform): Promise<boolean> {
+  const dirs = ["bat", "batcat"].map((n) => `${configHome(p, n)}/${n}`);
+  for (const dir of dirs) {
+    mkdirSync(dir, { recursive: true });
+    await Bun.write(`${dir}/config`, `# ${GENERATED}\n--theme="${BAT_THEME}"\n`);
+  }
+  return true;
+}
+
+/**
+ * delta, through git rather than a file of its own.
+ *
+ * delta has no config; it reads git config, which is where red-dev
+ * already told git to use it as the pager. Setting the syntax theme in
+ * the same place keeps one source of truth — and it reads bat's themes,
+ * so the same name serves both.
+ */
+export async function applyDelta(): Promise<boolean> {
+  if (!Bun.which("git")) return false;
+  const proc = Bun.spawn(["git", "config", "--global", "delta.syntax-theme", BAT_THEME], {
+    stdout: "ignore",
+    stderr: "ignore",
+    stdin: "ignore",
+  });
+  return (await proc.exited) === 0;
+}
+
+// ---------------------------------------------------------- lazygit
+
+/**
+ * lazygit, whose theme is a list of colour names or hexes.
+ *
+ * Written as its own file under a key lazygit merges, so anything else
+ * in the user's config survives — the same policy as zellij.
+ */
+export async function applyLazygit(p: Platform): Promise<boolean> {
+  if (!Bun.which("lazygit")) return false;
+  const dir = `${configHome(p, "lazygit")}/lazygit`;
+  mkdirSync(dir, { recursive: true });
+  const c = RED_DEV_ANSI;
+  const path = `${dir}/config.yml`;
+
+  // Merged by hand rather than parsed: a YAML parser is a dependency,
+  // and the only key we own is gui.theme. Everything above and below it
+  // is left exactly as found.
+  const block = `# ${GENERATED}
+gui:
+  theme:
+    activeBorderColor: ['${c.blue}', bold]
+    inactiveBorderColor: ['${c.brightBlack}']
+    selectedLineBgColor: ['${c.selectionBackground}']
+    optionsTextColor: ['${c.cyan}']
+    cherryPickedCommitBgColor: ['${c.selectionBackground}']
+    cherryPickedCommitFgColor: ['${c.green}']
+    unstagedChangesColor: ['${c.red}']
+`;
+  if (!existsSync(path)) {
+    await Bun.write(path, block);
+    return true;
+  }
+  const existing = await Bun.file(path).text();
+  // The prefix, not the whole marker: the wording changed with this
+  // file, and matching the full sentence would abandon every machine
+  // configured before it.
+  if (existing.includes(MARKER_PREFIX)) {
+    await Bun.write(path, block);
+    return true;
+  }
+  log.skip("lazygit config exists and is not ours — left alone");
+  return false;
+}
+
+// -------------------------------------------------- opencode, herdr
+
+/**
+ * opencode, following omarchy's lead.
+ *
+ * DHH sets `"theme": "system"` so the agent tracks whatever the terminal
+ * is doing rather than pinning a palette that will drift. Which is now
+ * exactly right: the terminal has one palette and it does not move.
+ */
+export async function applyOpencode(p: Platform): Promise<boolean> {
+  if (!Bun.which("opencode")) return false;
+  const dir = `${configHome(p, "opencode")}/opencode`;
+  mkdirSync(dir, { recursive: true });
+  const path = `${dir}/opencode.json`;
+
+  let cfg: Record<string, unknown> = {};
+  if (existsSync(path)) {
+    try {
+      cfg = JSON.parse(await Bun.file(path).text()) as Record<string, unknown>;
+    } catch {
+      log.skip("opencode.json is not valid JSON — left alone");
+      return false;
+    }
+  }
+  cfg["$schema"] = "https://opencode.ai/config.json";
+  cfg["theme"] = "system";
+  await Bun.write(path, `${JSON.stringify(cfg, null, 2)}\n`);
+  return true;
+}
+
+/**
+ * herdr, following the terminal.
+ *
+ * `name = "terminal"` makes herdr take the host terminal's ANSI palette
+ * rather than one of its own. Its theme.custom would accept sixteen
+ * hexes and we could emit them, but following cannot drift and sixteen
+ * copied values can.
+ *
+ * It used to prefer herdr's own theme where the name matched one of
+ * ours — tokyo-night, catppuccin, gruvbox, nord. Those names are gone,
+ * and so is the argument for them: there is no RedDB theme in herdr to
+ * prefer, and the terminal it would be disagreeing with is now fixed.
+ */
+export async function applyHerdr(p: Platform): Promise<boolean> {
+  if (!Bun.which("herdr") && !Bun.which("herdr.exe")) return false;
+
+  const { herdrConfigDir, herdrConfigPath } = await import("./herdr.ts");
+  const dir = herdrConfigDir(p);
+  const path = herdrConfigPath(p);
+  if (dir === null || path === null) return false;
+  mkdirSync(dir, { recursive: true });
+
+  const block = `# ${GENERATED}\n[theme]\nname = "terminal"\n`;
+
+  if (!existsSync(path)) {
+    await Bun.write(path, block);
+    return true;
+  }
+
+  const existing = await Bun.file(path).text();
+  // Only our own block is replaced. Keybindings, terminal settings and
+  // anything else in there are the user's — and herdr's config is where
+  // they live, so overwriting the file would take their keys with it.
+  const ours = /^# Generated by red-dev[^\n]*\n\[theme\]\nname = "[^"]*"\n/m;
+  if (ours.test(existing)) {
+    await Bun.write(path, existing.replace(ours, block));
+    return true;
+  }
+  if (/^\s*\[theme\]/m.test(existing)) {
+    log.skip("herdr already has a [theme] section of its own — left alone");
+    return false;
+  }
+  await Bun.write(path, `${existing.trimEnd()}\n\n${block}`);
+  return true;
+}
+
+// ------------------------------------------------------ the whole lot
+
+export interface TerminalSurfaceResult {
+  applied: string[];
+  skipped: string[];
+}
+
+/**
+ * Write the palette everywhere inside the terminal.
+ *
+ * Reasserted on every converge even though the content never changes.
+ * "Fixed" describes the bytes, not the frequency: a new machine, or one
+ * whose shared root was re-created, gets the palette because this runs,
+ * and a config edited by hand is put back the same way.
+ *
+ * Failure-tolerant per surface, deliberately. One unwritable config —
+ * herdr's directory on a machine where it was removed by hand, say —
+ * must not leave the other six untouched.
+ */
+export async function applyTerminalPalette(p: Platform): Promise<TerminalSurfaceResult> {
+  const writers: [string, () => Promise<boolean>][] = [
+    ["zellij", () => applyZellij(p)],
+    ["btop", () => applyBtop()],
+    ["bat", () => applyBat(p)],
+    ["delta", () => applyDelta()],
+    ["lazygit", () => applyLazygit(p)],
+    ["opencode", () => applyOpencode(p)],
+    ["herdr", () => applyHerdr(p)],
+  ];
+
+  const applied: string[] = [];
+  const skipped: string[] = [];
+
+  for (const [name, write] of writers) {
+    try {
+      if (await write()) applied.push(name);
+      else skipped.push(name);
+    } catch (err) {
+      log.warn(`${name}: ${(err as Error).message}`);
+      skipped.push(name);
+    }
+  }
+
+  return { applied, skipped };
+}
