+++ src/theme-apply.test.ts
@@ -29,22 +29,24 @@ const windows = {
 } satisfies Platform;
 
 describe("theme surfaces", () => {
-  test("native Windows receives the full portable CLI surface set", () => {
-    expect(themeSurfaceNames(windows)).toEqual([
-      "zellij",
-      "btop",
-      "neovim",
-      "vscode",
-      "bat",
-      "delta",
-      "lazygit",
-      "opencode",
-      "herdr",
-      "windows",
-    ]);
+  test("a theme reaches the desktop and nothing inside the terminal", () => {
+    // Nine of the eleven surfaces this used to name were terminal
+    // programs. They take the fixed palette now — see
+    // src/terminal-surfaces.ts and .red/adr/0002 — so a name reappearing
+    // in this list is a theme leaking back into a pane.
+    expect(themeSurfaceNames(windows)).toEqual(["vscode", "windows"]);
   });
 
-  test("native Windows stays in parity with WSL for portable CLI surfaces", () => {
+  test("no terminal program is a theme surface any more", () => {
+    const inTheTerminal = ["zellij", "btop", "neovim", "bat", "delta", "lazygit", "opencode", "herdr"];
+    for (const p of [windows, wsl]) {
+      for (const name of inTheTerminal) {
+        expect(themeSurfaceNames(p)).not.toContain(name);
+      }
+    }
+  });
+
+  test("native Windows stays in parity with WSL, apart from GNOME", () => {
     const portable = (name: string) => name !== "gnome";
 
     expect(themeSurfaceNames(windows).filter(portable)).toEqual(
