+++ src/theme-apply.ts
@@ -1,205 +1,29 @@
 /**
- * Apply a theme to every surface that can take one.
+ * Apply a theme to the surfaces a theme still owns.
  *
- * A theme in omakub is eight files, not a palette: alacritty, zellij,
- * btop, neovim, vscode, gnome, tophat and a wallpaper. Colouring only
- * the terminal is what makes a theme switch feel half-done — the
- * multiplexer keeps its old blue, the editor keeps its old background,
- * and the seams show immediately.
+ * It used to own eleven, nine of which were inside the terminal window.
+ * That is the arrangement .red/adr/0002 reverses: every program in a
+ * terminal paints over the ANSI slots beneath it, so a theme spread
+ * across nine of them arrived as neither the old one nor the new one,
+ * and switching looked like it had failed. The terminal now takes one
+ * fixed palette — src/terminal-surfaces.ts — and a theme is what changes
+ * the things you can see at a glance and nothing else overrides.
  *
- * Each writer here is independent and failure-tolerant: a missing
- * application is not an error, it is just a surface that does not exist
- * on this machine.
+ * What is left is the desktop: the wallpaper, the system accent and
+ * light-dark, and the editor that is a window rather than a pane.
+ *
+ * Each writer is independent and failure-tolerant: a missing application
+ * is not an error, it is just a surface that does not exist here.
  */
 
-import { existsSync, mkdirSync } from "node:fs";
 import { log } from "./log.ts";
-import { configHome } from "./shared-root.ts";
-import { applyBat, applyDelta, applyHerdr, applyLazygit, applyOpencode } from "./theme-cli.ts";
 import { applyWindowsDesktopTheme } from "./windows-theme.ts";
 import type { Platform } from "./platform.ts";
 import type { Theme } from "./themes.ts";
 
-function home(): string {
-  const h = process.env["HOME"] ?? process.env["USERPROFILE"];
-  if (!h) throw new Error("neither HOME nor USERPROFILE is set");
-  return h;
-}
-
-const GENERATED = "Generated by red-dev. Changes here are overwritten on theme switch.";
-
-// ----------------------------------------------------------- zellij
-
-/**
- * zellij reads themes from config.kdl or from a themes directory. We
- * write a themes/ file and reference it, so the user's own config.kdl
- * keeps whatever keybindings and layout settings they have set.
- */
-export async function applyZellij(theme: Theme, p: Platform): Promise<boolean> {
-  // configHome, not home(): when a shared root exists this writes into
-  // it, so the theme applied from WSL is the theme Windows reads. Every
-  // writer here used to hardcode ~/.config, which is what made the share
-  // something you migrated into rather than something configuration was
-  // born in.
-  //
-  // btop below deliberately keeps home(): its config names an absolute
-  // path to its own theme file, so a shared copy would point one side at
-  // the other's home.
-  const dir = `${configHome(p, "zellij")}/zellij`;
-  const themesDir = `${dir}/themes`;
-  mkdirSync(themesDir, { recursive: true });
-
-  const c = theme.terminal;
-  const kdl = `// ${GENERATED}
-themes {
-    red-dev {
-        fg "${c.foreground}"
-        bg "${c.background}"
-        black "${c.black}"
-        red "${c.red}"
-        green "${c.green}"
-        yellow "${c.yellow}"
-        blue "${c.blue}"
-        magenta "${c.purple}"
-        cyan "${c.cyan}"
-        white "${c.white}"
-        orange "${c.yellow}"
-    }
-}
-`;
-  await Bun.write(`${themesDir}/red-dev.kdl`, kdl);
-
-  // Point config.kdl at it without disturbing anything else in there.
-  const configPath = `${dir}/config.kdl`;
-  const existing = existsSync(configPath) ? await Bun.file(configPath).text() : "";
-
-  if (existing.includes('theme "red-dev"')) return true;
-
-  const withoutOldTheme = existing.replace(/^\s*theme\s+".*"\s*$/gm, "").trimEnd();
-  const next = `${withoutOldTheme}\n\n// ${GENERATED}\ntheme "red-dev"\n`;
-  await Bun.write(configPath, next.trimStart());
-  return true;
-}
-
-// ------------------------------------------------------------- btop
-
-export async function applyBtop(theme: Theme): Promise<boolean> {
-  const dir = `${home()}/.config/btop`;
-  const themesDir = `${dir}/themes`;
-  mkdirSync(themesDir, { recursive: true });
-
-  const c = theme.terminal;
-  // btop's format is flat key/value with quoted hex.
-  const btop = `# ${GENERATED}
-theme[main_bg]="${c.background}"
-theme[main_fg]="${c.foreground}"
-theme[title]="${c.foreground}"
-theme[hi_fg]="${c.blue}"
-theme[selected_bg]="${c.selectionBackground}"
-theme[selected_fg]="${c.foreground}"
-theme[inactive_fg]="${c.brightBlack}"
-theme[graph_text]="${c.foreground}"
-theme[proc_misc]="${c.green}"
-theme[cpu_box]="${c.blue}"
-theme[mem_box]="${c.green}"
-theme[net_box]="${c.purple}"
-theme[proc_box]="${c.cyan}"
-theme[div_line]="${c.brightBlack}"
-theme[temp_start]="${c.green}"
-theme[temp_mid]="${c.yellow}"
-theme[temp_end]="${c.red}"
-theme[cpu_start]="${c.green}"
-theme[cpu_mid]="${c.yellow}"
-theme[cpu_end]="${c.red}"
-theme[free_start]="${c.green}"
-theme[free_mid]="${c.cyan}"
-theme[free_end]="${c.blue}"
-theme[cached_start]="${c.cyan}"
-theme[cached_mid]="${c.blue}"
-theme[cached_end]="${c.purple}"
-theme[available_start]="${c.yellow}"
-theme[available_mid]="${c.yellow}"
-theme[available_end]="${c.red}"
-theme[used_start]="${c.green}"
-theme[used_mid]="${c.yellow}"
-theme[used_end]="${c.red}"
-theme[download_start]="${c.purple}"
-theme[download_mid]="${c.blue}"
-theme[download_end]="${c.cyan}"
-theme[upload_start]="${c.red}"
-theme[upload_mid]="${c.yellow}"
-theme[upload_end]="${c.green}"
-`;
-  await Bun.write(`${themesDir}/red-dev.theme`, btop);
-
-  // btop.conf is a flat file; rewrite only the color_theme line.
-  const confPath = `${dir}/btop.conf`;
-  const conf = existsSync(confPath) ? await Bun.file(confPath).text() : "";
-  const line = `color_theme = "${themesDir}/red-dev.theme"`;
-
-  if (conf.includes("color_theme")) {
-    await Bun.write(confPath, conf.replace(/^color_theme\s*=.*$/m, line));
-  } else {
-    await Bun.write(confPath, `${conf.trimEnd()}\n${line}\ntheme_background = False\n`.trimStart());
-  }
-  return true;
-}
-
-// ----------------------------------------------------------- neovim
-
-/**
- * LazyVim loads every spec under lua/plugins, so dropping one file in
- * both declares the colorscheme plugin and selects it. Writing into
- * LazyVim's own files instead would be overwritten by its updater.
- */
-export async function applyNeovim(theme: Theme): Promise<boolean> {
-  const pluginsDir = `${home()}/.config/nvim/lua/plugins`;
-  if (!existsSync(`${home()}/.config/nvim`)) return false;
-  mkdirSync(pluginsDir, { recursive: true });
-
-  // Each colorscheme needs its plugin declared, or LazyVim silently
-  // falls back and the theme switch looks like it did nothing.
-  const PLUGIN: Record<string, string> = {
-    tokyonight: '"folke/tokyonight.nvim"',
-    "catppuccin-macchiato": '"catppuccin/nvim", name = "catppuccin"',
-    gruvbox: '"ellisonleao/gruvbox.nvim"',
-    everforest: '"neanias/everforest-nvim"',
-    kanagawa: '"rebelot/kanagawa.nvim"',
-    matteblack: '"tahayvr/matteblack.nvim"',
-    nordfox: '"EdenEast/nightfox.nvim"',
-    bamboo: '"ribru17/bamboo.nvim"',
-    "monokai-pro": '"gthelding/monokai-pro.nvim"',
-    "rose-pine-dawn": '"rose-pine/neovim", name = "rose-pine"',
-  };
-  const spec = PLUGIN[theme.neovim];
-  if (!spec) throw new Error(`no Neovim plugin mapped for ${theme.neovim}`);
-
-  const lua = `-- ${GENERATED}
-return {
-  { ${spec}, lazy = false, priority = 1000 },
-  {
-    "LazyVim/LazyVim",
-    opts = { colorscheme = "${theme.neovim}" },
-  },
-}
-`;
-  await Bun.write(`${pluginsDir}/red-dev-theme.lua`, lua);
-  return true;
-}
-
 // --------------------------------------------------------- entrypoint
 
-const PORTABLE_SURFACE_NAMES = [
-  "zellij",
-  "btop",
-  "neovim",
-  "vscode",
-  "bat",
-  "delta",
-  "lazygit",
-  "opencode",
-  "herdr",
-];
+const PORTABLE_SURFACE_NAMES = ["vscode"];
 
 export function themeSurfaceNames(p: Platform): string[] {
   return [
@@ -228,15 +52,7 @@ export async function applyThemeEverywhere(
   const key = slug ?? theme.name.toLowerCase().replace(/\s+/g, "-");
 
   const surfaceFns: Record<string, () => Promise<boolean>> = {
-    zellij: () => applyZellij(theme, p),
-    btop: () => applyBtop(theme),
-    neovim: () => applyNeovim(theme),
     vscode: () => applyVsCodeTheme(theme, p, key),
-    bat: () => applyBat(theme, p, key),
-    delta: () => applyDelta(theme, key),
-    lazygit: () => applyLazygit(theme, p),
-    opencode: () => applyOpencode(p),
-    herdr: () => applyHerdr(p, key),
     windows: () => applyWindowsDesktopTheme(theme, p, key),
     gnome: () => applyGnomeTheme(p, theme, key),
   };
