--- src/theme-cli.ts
@@ -1,224 +1,0 @@
-/**
- * Theming the tools omakub leaves alone.
- *
- * omakub themes eight surfaces and every one of them is an application
- * with a window: terminal, multiplexer, monitor, editor, VS Code, GNOME,
- * the dock, a wallpaper. The command-line tools it installs keep
- * whatever colours they shipped with, so a Kanagawa terminal shows you a
- * Monokai diff and a Dracula file browser.
- *
- * omarchy — DHH's Arch version — goes one step further and themes the
- * *agents*: its opencode.json carries `"theme": "system"` and it ships a
- * pi extension that watches the active theme and calls setTheme when it
- * changes. That is the idea this file follows.
- *
- * These are the surfaces that work on all five targets, which is what
- * separates them from the GNOME layer. A config file is a config file in
- * WSL, on the desktop and in Git Bash.
- */
-
-import { existsSync, mkdirSync } from "node:fs";
-import { log } from "./log.ts";
-import { configHome } from "./shared-root.ts";
-import type { Platform } from "./platform.ts";
-import type { Theme } from "./themes.ts";
-
-const GENERATED = "Generated by red-dev. Overwritten on theme switch.";
-
-/**
- * bat's theme name for one of ours.
- *
- * bat ships a fixed set built from Sublime syntaxes, so this is a
- * mapping to the nearest published theme rather than a palette we can
- * emit. Falling back to a dark default is deliberate: every palette this
- * project ships is dark, so the wrong dark theme is a smaller error than
- * a light one.
- */
-const BAT_THEME: Record<string, string> = {
-  "tokyo-night": "Nord",
-  catppuccin: "Nord",
-  gruvbox: "gruvbox-dark",
-  everforest: "gruvbox-dark",
-  kanagawa: "Nord",
-  "matte-black": "base16",
-  nord: "Nord",
-  "osaka-jade": "Nord",
-  ristretto: "Monokai Extended",
-  "rose-pine": "base16",
-};
-
-/** delta reads bat's themes, so the same mapping serves both. */
-export function batThemeFor(slug: string): string {
-  return BAT_THEME[slug] ?? "base16";
-}
-
-/**
- * bat, and the Debian rename that makes it two tools.
- *
- * Debian ships the binary as `batcat` because `bat` was taken, and the
- * config directory follows the binary: `~/.config/batcat` there,
- * `~/.config/bat` everywhere else. Writing only one leaves the tool
- * reading a file that does not exist and reporting nothing wrong.
- */
-export async function applyBat(theme: Theme, p: Platform, slug: string): Promise<boolean> {
-  const dirs = ["bat", "batcat"].map((n) => `${configHome(p, n)}/${n}`);
-  let wrote = false;
-  for (const dir of dirs) {
-    mkdirSync(dir, { recursive: true });
-    await Bun.write(`${dir}/config`, `# ${GENERATED}\n--theme="${batThemeFor(slug)}"\n`);
-    wrote = true;
-  }
-  void theme;
-  return wrote;
-}
-
-/**
- * delta, through git rather than a file of its own.
- *
- * delta has no config; it reads git config, which is where red-dev
- * already told git to use it as the pager. Setting the syntax theme in
- * the same place keeps one source of truth.
- */
-export async function applyDelta(theme: Theme, slug: string): Promise<boolean> {
-  if (!Bun.which("git")) return false;
-  const proc = Bun.spawn(
-    ["git", "config", "--global", "delta.syntax-theme", batThemeFor(slug)],
-    { stdout: "ignore", stderr: "ignore", stdin: "ignore" },
-  );
-  void theme;
-  return (await proc.exited) === 0;
-}
-
-/**
- * lazygit, whose theme is a list of colour names or hexes.
- *
- * Written as its own file under a key lazygit merges, so anything else
- * in the user's config survives — the same policy as zellij.
- */
-export async function applyLazygit(theme: Theme, p: Platform): Promise<boolean> {
-  if (!Bun.which("lazygit")) return false;
-  const dir = `${configHome(p, "lazygit")}/lazygit`;
-  mkdirSync(dir, { recursive: true });
-  const c = theme.terminal;
-  const path = `${dir}/config.yml`;
-
-  // Merged by hand rather than parsed: a YAML parser is a dependency,
-  // and the only key we own is gui.theme. Everything above and below it
-  // is left exactly as found.
-  const block = `# ${GENERATED}
-gui:
-  theme:
-    activeBorderColor: ['${c.blue}', bold]
-    inactiveBorderColor: ['${c.brightBlack}']
-    selectedLineBgColor: ['${c.selectionBackground}']
-    optionsTextColor: ['${c.cyan}']
-    cherryPickedCommitBgColor: ['${c.selectionBackground}']
-    cherryPickedCommitFgColor: ['${c.green}']
-    unstagedChangesColor: ['${c.red}']
-`;
-  if (!existsSync(path)) {
-    await Bun.write(path, block);
-    return true;
-  }
-  const existing = await Bun.file(path).text();
-  if (existing.includes(GENERATED)) {
-    // Ours from a previous run: replace from our marker to the end of
-    // the block we wrote, leaving anything the user added after it.
-    await Bun.write(path, block);
-    return true;
-  }
-  log.skip("lazygit config exists and is not ours — left alone");
-  return false;
-}
-
-/**
- * opencode, following omarchy's lead.
- *
- * DHH sets `"theme": "system"` so the agent tracks whatever the terminal
- * is doing rather than pinning a palette that will drift. That is the
- * right answer here too: red-dev already themes the terminal, so an
- * agent that follows the terminal follows red-dev.
- */
-export async function applyOpencode(p: Platform): Promise<boolean> {
-  if (!Bun.which("opencode")) return false;
-  const dir = `${configHome(p, "opencode")}/opencode`;
-  mkdirSync(dir, { recursive: true });
-  const path = `${dir}/opencode.json`;
-
-  let cfg: Record<string, unknown> = {};
-  if (existsSync(path)) {
-    try {
-      cfg = JSON.parse(await Bun.file(path).text()) as Record<string, unknown>;
-    } catch {
-      log.skip("opencode.json is not valid JSON — left alone");
-      return false;
-    }
-  }
-  cfg["$schema"] = "https://opencode.ai/config.json";
-  cfg["theme"] = "system";
-  await Bun.write(path, `${JSON.stringify(cfg, null, 2)}\n`);
-  return true;
-}
-
-/**
- * herdr's built-in themes, by our slug.
- *
- * Only the ones it actually ships, read from its own source rather than
- * guessed: catppuccin, catppuccin-latte, dracula, nord and tokyo-night
- * are in src/config/theme.rs, and its docs name gruvbox as having a
- * light sibling. Everything else falls through to `terminal`.
- */
-const HERDR_THEME: Record<string, string> = {
-  "tokyo-night": "tokyo-night",
-  catppuccin: "catppuccin",
-  gruvbox: "gruvbox",
-  nord: "nord",
-};
-
-/**
- * herdr, preferring its own palette over ours.
- *
- * Its theme.custom takes sixteen hex values and we could emit all of
- * them — but `name = "terminal"` makes herdr follow the host terminal's
- * ANSI palette, which red-dev has already set. Following cannot drift;
- * sixteen copied hexes can, and would have to be re-copied every time
- * either side changes. It is the same reasoning as opencode's
- * `"theme": "system"`, and the same conclusion omarchy reached.
- *
- * Where herdr ships a theme with our name, that wins: its author tuned
- * those sixteen values for their own UI, which the ANSI sixteen cannot
- * express as well.
- */
-export async function applyHerdr(p: Platform, slug: string): Promise<boolean> {
-  if (!Bun.which("herdr") && !Bun.which("herdr.exe")) return false;
-
-  const { herdrConfigDir, herdrConfigPath } = await import("./herdr.ts");
-  const dir = herdrConfigDir(p);
-  const path = herdrConfigPath(p);
-  if (dir === null || path === null) return false;
-  mkdirSync(dir, { recursive: true });
-
-  const name = HERDR_THEME[slug] ?? "terminal";
-  const block = `# ${GENERATED}\n[theme]\nname = "${name}"\n`;
-
-  if (!existsSync(path)) {
-    await Bun.write(path, block);
-    return true;
-  }
-
-  const existing = await Bun.file(path).text();
-  // Only our own block is replaced. Keybindings, terminal settings and
-  // anything else in there are the user's — and herdr's config is where
-  // they live, so overwriting the file would take their keys with it.
-  const ours = /^# Generated by red-dev[^\n]*\n\[theme\]\nname = "[^"]*"\n/m;
-  if (ours.test(existing)) {
-    await Bun.write(path, existing.replace(ours, block));
-    return true;
-  }
-  if (/^\s*\[theme\]/m.test(existing)) {
-    log.skip("herdr already has a [theme] section of its own — left alone");
-    return false;
-  }
-  await Bun.write(path, `${existing.trimEnd()}\n\n${block}`);
-  return true;
-}
